+++ source/convert_multichannel_audio_to_aac_or_ac3/changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.4</span>**
+- add option to keep original multichannel sources
+- fix error message text to simply refer to multichannel streams instead of DTS streams if no streams are found to encode (consistent with 0.0.2 fix)
+
 **<span style="color:#56adda">0.0.3</span>**
 - copy-paste error